separate get item function and it can be use for fetching top stories each item id

+++ client/fetch_item.go
@@ -19,7 +19,7 @@ func DisplayItem(output string, typeOfPost string, itemURL string) {
 }
 
 // GetItem function returns the item that related ID
-func GetItem(itemID int) {
+func GetItem(itemID int64) {
 	var ItemResponse model.Item
 	itemURL := fmt.Sprintf("%s%d%s", BaseURL, itemID, ".json")
 	item, err := http.Get(itemURL)
